+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Enhanced file management and analysis capabilities (supports xls, csv, and pdf files)
 - Improved MariaDB improvements
 - Add support for loading up to 60K tables when connecting data sources
+- Added automated tests for postgres database
 
 ## Version 0.0.263 (December 4, 2025)
 - System prompt improvements and a new section for analytical standards